- Fixed compilation error with Workout filename - Moved WorkoutController one folder up for consistency
